Moved mention plugin related code
to a separate directory.
