+++ src/components/editor-input/mention-plugin/mention-component.jsx
