+++ src/components/editor-input/mention-plugin/mention-node.jsx
