+++ src/components/editor-input/mention-plugin/option-item.js
